+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Enhancement
 
 * `Realm.compactRealm()` works for encrypted Realms.
+*  Added `first(E defaultValue)` and `last(E defaultValue)` methods to `RealmList` and `RealmResult`. These methods will return the provided object instead of throwing an `IndexOutOfBoundsException` if the list is empty.
 *  Added `User.all()` that returns all known Realm Object Server users.
 
 ## 2.0.3